refactor: subclass Linkindexer to add hashing functionality
